requirements-test.txt
boto3>=1.28.0
pandas>=2.0.0
seaborn>=0.12.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
numpy>=1.24.0


